neutrino.inc: Try to improve handling of update.urls files
This is an attempt to generate the update URLs for the image update.
It has not been tested yet and is not yet actively usable because some
things still need to be adjusted. For example, it is important to adjust
the local configuration in local.common.conf, which must also be
implemented in the buildenv script. Adjustments to the tuxbox.conf are
also required. The generation is basically created so that the address
through which an image file can be accessed points to the deploy
directories located within the build directory of the respective machine,
e.g.: build/<machine>/tmp/deploy/images/<machinebuild>/. These changes
also make it necessary to adjust the feeds for the ipks. The buildenv
script must also take this into account when generating the dist directory.